requirements.txt
opencv-python~=3.4.2.16
numpy~=1.18.1
fire~=0.3.1
matplotlib~=3.1.3

pillow~=7.1.2
scipy~=1.4.1